requirements.txt
fastapi
httpx
pydantic
pydantic_settings
pyinstaller
python-dotenv
Requests
starlette
uvicorn